% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{ph_audio_discard}
\alias{ph_audio_discard}
\title{Discard a visit's audio.}
\usage{
ph_audio_discard(part)
}
\arguments{
\item{part}{A path from \code{\link[=ph_audio_open]{ph_audio_open()}}.}
}
\value{
\code{TRUE} if a file was removed.
}
\description{
Discard a visit's audio.
}
\examples{
\dontrun{
ph_audio_discard(p)
}
}
