% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic.R
\name{ph_browser_name}
\alias{ph_browser_name}
\title{Which browser is this?}
\usage{
ph_browser_name(ua)
}
\arguments{
\item{ua}{A \code{navigator.userAgent} string.}
}
\value{
A short browser name, or \code{NA_character_} if it is not recognised.
}
\description{
Which browser is this?
}
\examples{
ph_browser_name("Mozilla/5.0 (Macintosh) Gecko/20100101 Firefox/141.0")
}
