% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{ph_cli}
\alias{ph_cli}
\title{Command line entry point.}
\usage{
ph_cli(args = character(0))
}
\arguments{
\item{args}{Character vector of arguments, as from \code{commandArgs()}.}
}
\value{
Invisibly, the command's return value.
}
\description{
Command line entry point.
}
\examples{
\dontrun{
ph_cli(c("status", "--work", "~/phostor/family"))
}
}
