% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{ph_is_web_format}
\alias{ph_is_web_format}
\title{Is this a format a browser can display directly?}
\usage{
ph_is_web_format(path)
}
\arguments{
\item{path}{A file path.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is this a format a browser can display directly?
}
\examples{
ph_is_web_format("a.jpg")
ph_is_web_format("a.heic")
}
