% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{ph_path_read}
\alias{ph_path_read}
\title{Read a session's path.}
\usage{
ph_path_read(session_dir)
}
\arguments{
\item{session_dir}{A session directory.}
}
\value{
A data.frame with the columns of \code{path.tsv}; empty if there is none.
}
\description{
Read a session's path.
}
\examples{
\dontrun{
ph_path_read(d)
}
}
